+++ CHANGELOG.md
@@ -14,6 +14,7 @@
     * Support Pooled opcode budget
     * Txnas, Gtxnas, Gtxnsas, Args, Log (logs are stored in txReceipt)
 * Update all transaction functions (eg. `executeTx`, `addAsset`, `addApp` ..etc) to return a transaction receipt. Add `runtime.getTransactionInfo` in `@algo-builder/runtime` to query transaction info.
++ Add Asset Name to `assetDefinition` in `@algo-builder/runtime`.
 
 ### Breaking changes
 `@algo-builder/runtime`: